Fix an upload retryability bug
Before this change, if the upload failed because of an empty file, then
a retryable=false exception was thrown.

And then it was retried anyway.

With this change in place, if we wrap one UploadException in another,
we will correctly ask the wrapped exception for its retryability status
rather than just assuming "true".